Rename _elm_common to elm_common to satisfy pylint import-private-name
